fix(cli): use stable MCP client identities and explicit config edits
